refactor: reorganize tags module into package structure
- Split monolithic tags.py into modular package structure
- Create separate modules for base, special, stock tags, and SVG
- Extract configuration (HTML_ATTRIBUTES) into config.py
- Move utility functions (SafeStr, clean_html_attr_key, locals_cleanup) to utils.py
- Improve code organization with clear separation of concerns
- Update imports in __init__.py, responses.py, and templates.py
- Maintain backward compatibility with existing API

This refactoring improves maintainability and makes the codebase more modular
without changing any external functionality.
