+++ src/air/tags/models/stock.py
@@ -1,595 +1,7 @@
-"""Easy to write and performant HTML content generation using Python classes to render HTML."""
-
-import html
-from functools import cached_property
 from typing import Any
 
-
-def clean_html_attr_key(key: str) -> str:
-    """Clean up HTML attribute keys to match the standard W3C HTML spec.
-
-    Args:
-        key: An uncleaned HTML attribute key
-
-    Returns:
-
-        Cleaned HTML attribute key
-    """
-    # If a "_"-suffixed proxy for "class", "for", or "id" is used,
-    # convert it to its normal HTML equivalent.
-    key = dict(class_="class", for_="for", id_="id", as_="as").get(key, key)
-    # Remove leading underscores and replace underscores with dashes
-    return key.lstrip("_").replace("_", "-")
-
-
-class Tag:
-    """Base tag for all other tags.
-
-    Sets four attributes, name, module, children, and attrs.
-    These are important for Starlette view responses, as nested objects
-    get auto-serialized to JSON and need to be rebuilt. With
-    the values of these attributes, the object reconstruction can occur.
-    """
-
-    self_closing = False
-
-    def __init__(self, *children: Any, **kwargs: str | int | float | bool):
-        """
-        Args:
-            children: Tags, strings, or other rendered content.
-            kwargs: Keyword arguments transformed into tag attributes.
-        """
-        self._name = self.__class__.__name__
-        self._module = self.__class__.__module__
-        self._children, self._attrs = children, kwargs
-
-    @property
-    def name(self) -> str:
-        return self._name.lower()
-
-    @property
-    def attrs(self) -> str:
-        if not self._attrs:
-            return ""
-        attrs = []
-        for k, v in self._attrs.items():
-            if isinstance(v, bool) and v is True:
-                # Add single word attribute like "selected"
-                attrs.append(k)
-            elif isinstance(v, bool) and v is False:
-                # Skip single word attribute like "selected"
-                continue
-            else:
-                attrs.append(f'{clean_html_attr_key(k)}="{v}"')
-        return " " + " ".join(attrs)
-
-    @cached_property
-    def children(self):
-        elements = []
-        for child in self._children:
-            if isinstance(child, Tag):
-                elements.append(child.render())
-            elif isinstance(child, SafeStr):
-                elements.append(child)
-            elif isinstance(child, str):
-                elements.append(html.escape(child))
-            else:
-                # If the type isn't supported, we just convert to `str`
-                # and then escape it for safety. This matches to what most
-                # template tools do, which prevents hard bugs in production
-                # from stopping users cold.
-                elements.append(html.escape(str(child)))
-        return "".join(elements)
-
-    def render(self) -> str:
-        if self.name == "tag":
-            return self.children
-        if self.self_closing:
-            return f"<{self.name}{self.attrs} />"
-        return f"<{self.name}{self.attrs}>{self.children}</{self.name}>"
-
-
-class CaseTag(Tag):
-    """This is for case-sensitive tags like those used in SVG generation."""
-
-    @property
-    def name(self) -> str:
-        return self._name[0].lower() + self._name[1:]
-
-
-# Utilities
-
-
-class SafeStr(str):
-    """A string subclass that doesn't trigger html.escape() when called by Tag.render()
-
-    Example:
-        sample = SafeStr('Hello, world')
-    """
-
-    def __new__(cls, value):
-        obj = super().__new__(cls, value)
-        return obj
-
-    def __repr__(self):
-        return super().__repr__()
-
-
-# Special tags
-
-
-class Html(Tag):
-    """Defines the root of an HTML document"""
-
-    def render(self) -> str:
-        return f"""<!doctype html><html{self.attrs}>{self.children}</html>"""
-
-
-class Raw(Tag):
-    """Renders raw HTML content without escaping.
-
-    Args:
-        args: A single string containing raw text to render
-
-    Raises:
-        TypeError: If non-string content is provided
-        ValueError: If multiple arguments are provided
-
-    Example:
-
-        # Produces '<strong>Bold</strong> text'
-        Raw('<strong>Bold</strong> text')
-
-        # Use with other tags
-        Div(
-            P("Safe content"),
-            Raw('<hr class="divider">'),
-            P("More safe content")
-        )
-    """
-
-    def __init__(self, *args: Any, **kwargs: str | int | float | bool):
-        """Initialize Raw with a single string argument.
-
-        Args:
-            *args: Should be exactly one string argument
-            **kwargs: Ignored (for consistency with Tag interface)
-        """
-        if len(args) > 1:
-            raise ValueError("Raw accepts only one string argument")
-
-        raw_string: str = args[0] if args else ""
-
-        if not isinstance(raw_string, str):
-            raise TypeError("Raw only accepts string content")
-
-        super().__init__(raw_string)
-
-    def render(self) -> str:
-        """Render the string without escaping."""
-        return self._children[0] if self._children else ""
-
-
-class NoEscapeTag(Tag):
-    """Custom tag that does not escape its children.
-
-    This is used for tags like Script and Style where content
-    should not be HTML-escaped.
-    """
-
-    def render(self) -> str:
-        """Render the tag with unescaped content."""
-        content = self._children[0] if self._children else ""
-        return f"<{self.name}{self.attrs}>{content}</{self.name}>"
-
-
-class Script(NoEscapeTag):
-    """Defines a client-side script
-
-    Warning: Script tag does not protect against code injection.
-    """
-
-    pass
-
-
-class Style(NoEscapeTag):
-    """Defines style information for a document
-
-    Warning: Style tag does not protect against code injection.
-    """
-
-    pass
-
-
-class Children(Tag):
-    def render(self) -> str:
-        return self.children
-
-
-class Tags(Tag):
-    def render(self) -> str:
-        return self.children
-
-
-# HTML tag attribute map
-
-html_attributes = {
-    "A": [
-        "href",
-        "target",
-        "download",
-        "rel",
-        "hreflang",
-        "typereferrerpolicy",
-        "media",
-        "ping",
-        "class_",
-        "id",
-    ],
-    "Area": [
-        "alt",
-        "coords",
-        "download",
-        "href",
-        "ping",
-        "referrerpolicy",
-        "rel",
-        "shape",
-        "target",
-    ],
-    "Audio": [
-        "autoplay",
-        "controls",
-        "loop",
-        "muted",
-        "preload",
-        "src",
-    ],
-    "Base": [
-        "href",
-        "target",
-    ],
-    "Bdi": [
-        "dir",
-    ],
-    "Blockquote": [
-        "cite",
-    ],
-    "Button": [
-        "name",
-        "type",
-        "value",
-        "autofocus",
-        "disabled",
-        "form",
-        "formaction",
-        "formenctype",
-        "formmethod",
-        "formnovalidate",
-        "formtarget",
-        "popovertarget",
-        "popovertargetaction",
-    ],
-    "Canvas": [
-        "width",
-        "height",
-    ],
-    "Col": ["span"],
-    "Colgroup": ["span"],
-    "Data": ["value"],
-    "Dd": ["cite", "datetime"],
-    "Details": ["open"],
-    "Dialog": ["open"],
-    "Embed": [
-        "src",
-        "typewidthheight",
-    ],
-    "Fieldset": [
-        "disabled",
-        "form",
-        "name",
-    ],
-    "Form": [
-        "accept-charset",
-        "action",
-        "autocomplete",
-        "enctype",
-        "method",
-        "name",
-        "novalidate",
-        "rel",
-        "target",
-    ],
-    "Head": [
-        "profile",
-    ],
-    "Iframe": [
-        "src",
-        "srcdoc",
-        "width",
-        "height",
-        "allow",
-        "allowfullscreen",
-        "allowpaymentrequest",
-        "loading",
-        "name",
-        "referrerpolicy",
-        "sandbox",
-    ],
-    "Img": [
-        "src",
-        "width",
-        "heightsrcset",
-        "alt",
-        "crossorigin",
-        "ismap",
-        "loading",
-        "longdesc",
-        "referrerpolicy",
-        "sizesusemap",
-    ],
-    "Input": [
-        "type",
-        "value",
-        "readonly",
-        "required",
-        "accept",
-        "alt",
-        "autocomplete",
-        "autofocus",
-        "checked",
-        "dirname",
-        "disabled",
-        "form",
-        "formaction",
-        "formenctype",
-        "formmethod",
-        "formnovalidate",
-        "formtarget",
-        "height",
-        "list",
-        "max",
-        "maxlength",
-        "min",
-        "minlength",
-        "multiple",
-        "name",
-        "pattern",
-        "placeholder",
-        "popovertarget",
-        "popovertargetaction",
-        "size",
-        "src",
-        "step",
-    ],
-    "Ins": [
-        "cite",
-        "datetime",
-    ],
-    "Label": ["for"],
-    "Li": [
-        "value",
-    ],
-    "Link": [
-        "href",
-        "as_",
-        "blocking",
-        "crossorigin",
-        "disabled",
-        "fetchpriority",
-        "hreflang",
-        "imagesizes",
-        "imagesrcset",
-        "integrity",
-        "media",
-        "referrerpolicy",
-        "rel",
-        "sizes",
-        "title",
-        "type",
-    ],
-    "Map": [
-        "name",
-    ],
-    "Marquee": [
-        "behavior",
-        "direction",
-        "height",
-        "width",
-        "loop",
-        "scrollamount",
-        "scrolldelay",
-        "truespeed",
-        "vspace",
-    ],
-    "Menu": [
-        "compact",
-    ],
-    "Meta": [
-        "charset",
-        "content",
-        "http_equiv",
-        "media",
-        "name",
-    ],
-    "Meter": [
-        "value",
-        "min",
-        "max",
-        "low",
-        "high",
-        "optimum",
-    ],
-    "Object": [
-        "archive",
-        "border",
-        "classidcodebase",
-        "codetype",
-        "data",
-        "declare",
-        "form",
-        "height",
-        "name",
-        "standby",
-        "type",
-        "usemap",
-        "width",
-    ],
-    "Ol": [
-        "compact",
-        "reversed",
-        "start",
-        "type",
-    ],
-    "Optgroup": [
-        "disabled",
-        "label",
-    ],
-    "Option": [
-        "disabled",
-        "label",
-        "selected",
-        "value",
-    ],
-    "Output": [
-        "for_",
-        "form",
-        "name",
-    ],
-    "Pre": [
-        "width",
-        "wrap",
-    ],
-    "Progress": [
-        "max",
-        "value",
-    ],
-    "Q": [
-        "cite",
-    ],
-    "Script": [
-        "async",
-        "attributionsrc",
-        "blocking",
-        "crossorigin",
-        "defer",
-        "fetchpriority",
-        "integrity",
-        "nomodule",
-        "noncereferrerpolicy",
-        "src",
-        "type",
-    ],
-    "Select": [
-        "autocomplete",
-        "autofocus",
-        "disabled",
-        "form",
-        "multiple",
-        "name",
-        "required",
-        "size",
-    ],
-    "Slot": [
-        "name",
-    ],
-    "Source": [
-        "src",
-        "type",
-        "sizes",
-        "media",
-        "srcset",
-        "height",
-        "width",
-    ],
-    "Style": [
-        "media",
-        "nonce",
-        "title",
-        "blocking",
-    ],
-    "Td": [
-        "colspan",
-        "rowspan",
-        "headers",
-    ],
-    "Template": [
-        "shadowrootmode",
-        "shadowrootdelegatesfocus",
-        "shadowrootclonable",
-        "shadowrootserializable",
-    ],
-    "Textarea": [
-        "autocapitalize",
-        "autocomplete",
-        "autocorrect",
-        "autofocus",
-        "cols",
-        "dirname",
-        "disabled",
-        "form",
-        "maxlength",
-        "minlength",
-        "name",
-        "placeholder",
-        "readonly",
-        "required",
-        "rows",
-        "spellcheck",
-        "wrap",
-    ],
-    "Th": [
-        "abbr",
-        "colspan",
-        "headers",
-        "rowspan",
-        "scope",
-    ],
-    "Time": [
-        "datetime",
-    ],
-    "Track": [
-        "default",
-        "kind",
-        "label",
-        "srclang",
-        "src",
-    ],
-    "Ul": [
-        "compact",
-        "type",
-    ],
-    "Video": [
-        "src",
-        "autoplay",
-        "controls",
-        "controlslist",
-        "crossorigin",
-        "disablepictureinpicture",
-        "disableremoteplayback",
-        "height",
-        "width",
-        "loop",
-        "muted",
-        "playsinline",
-        "poster",
-        "preload",
-    ],
-}
-
-
-def locals_cleanup(local_data, obj):
-    """Converts arguments to kwargs per the html_attributes structure"""
-    data = {}
-    attrs = html_attributes.get(obj.__class__.__name__, [])
-    attrs += ["class_", "id", "style"]
-    for attr in attrs:
-        if local_data.get(attr) is not None:
-            data[attr] = local_data[attr]
-    return data
-
-
-# Stock tags
+from ..utils import locals_cleanup
+from .base import Tag
 
 
 class A(Tag):
